feat: added query safety test to customSQLTest
