+++ ingestion/tests/unit/test_helpers.py
@@ -24,6 +24,7 @@ from metadata.generated.schema.type.tagLabel import (
 from metadata.utils.helpers import (
     clean_up_starting_ending_double_quotes_in_string,
     get_entity_tier_from_tags,
+    is_safe_sql_query,
     list_to_dict,
 )
 
@@ -85,3 +86,39 @@ class TestHelpers(TestCase):
         )
 
         assert get_entity_tier_from_tags(table_entity_wo_tier.tags) is None
+
+    def test_is_safe_sql_query(self):
+        """Test is_safe_sql_query function"""
+
+        delete_query = """
+        DELETE FROM airflow_task_instance
+        WHERE dag_id = 'test_dag_id'
+        """
+
+        drop_query = """
+        DROP TABLE IF EXISTS test_table
+        """
+
+        create_query = """
+        CREATE TABLE test_table (
+            id INT,
+            name VARCHAR(255)
+        )
+        """
+
+        select_query = """
+        SELECT * FROM test_table
+        """
+
+        cte_query = """
+        WITH foo AS (
+            SELECT * FROM test_table
+        )
+        SELECT * FROM foo
+        """
+
+        self.assertFalse(is_safe_sql_query(delete_query))
+        self.assertFalse(is_safe_sql_query(drop_query))
+        self.assertFalse(is_safe_sql_query(create_query))
+        self.assertTrue(is_safe_sql_query(select_query))
+        self.assertTrue(is_safe_sql_query(cte_query))
