+++ ingestion/tests/unit/test_suite/test_validations_databases.py
@@ -225,6 +225,12 @@ EXECUTION_DATE = datetime.strptime("2021-07-03", "%Y-%m-%d")
             "TABLE",
             (TestCaseResult, "6", None, TestCaseStatus.Success),
         ),
+        (
+            "test_case_unsafe_table_custom_sql_query",
+            "tableCustomSQLQuery",
+            "TABLE",
+            (TestCaseResult, None, None, TestCaseStatus.Aborted),
+        ),
     ],
 )
 def test_suite_validation_database(
